requirements.txt
pyTelegramBotAPI==4.14.0
psutil==5.9.6
Pillow==10.0.0
pycryptodome==3.19.0
pyinstaller==5.13.2
requests==2.31.0